Redoes button functionality
Previous functionality was achieved by checking if previous button click
was a number or not, and if current button click is a number or not.
This commit changed the functionality so that it's only based on what
the current button click is.

This also means that non-numeric values are no longer displayed on the
display field.