Changes CE button to AC
This is a cosmetic change, the function of the button is the exact same.
CE usually means "Clear Entry", so in the future is will be used to
remove the last entered entry into the calculator. AC means "All Clear,"
and therefore is better for clearing all currently entered values of the
calculator.